In my draft html pages, I can now click on the view button and view the project details, name and descriptions(the list of project members will come later). This should complete US8. Try connecting it to the angular frontend.

+++ conduite/src/main/java/com/example/conduite/repos/AppUserRepo.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.conduite.entities.AppUser;
 
 @Repository
-public interface AppUserRepo extends JpaRepository<AppUser, Integer> {
+public interface AppUserRepo extends JpaRepository<AppUser, Long> {
 }
